feat(explorer): put the ranking table on the compare route and make /results an overview
The compare route's empty state was a heading and a link to the run picker.
The ranking table it linked past - the filters, the cross-benchmark matrix -
is what a reader arriving at "compare" is looking for, so the table now IS
that state. A single pinned run lands there too: one run is not a comparison,
and the tray already carries the pin.

That leaves /results as the corpus overview it had buried below the fold:
what the corpus holds, what arrived most recently, and where to browse. It
loads unfiltered, because a page answering "what is in here" should not open
behind a filter. Its title and nav entry are now Overview, which is what it
shows.

The ranking half moves to Leaderboard.tsx unchanged apart from its name and a
notice slot for the compare route's messages. Home.tsx keeps the overview
half. The two suites read one set of corpus fixtures so they cannot drift.
