+++ results-explorer/src/pages/Home.tsx
@@ -1,133 +1,50 @@
-import { useEffect, useMemo, useRef, useState } from "preact/hooks";
+import { useEffect, useRef, useState } from "preact/hooks";
 import type { RoutableProps } from "preact-router";
-import type {
-  MetaCohort,
-  MetaCohortPlatform,
-  MetaLeaderboard as MetaLeaderboardData,
-  MetaPlatform,
-  MetaRank,
-} from "@/types";
+import type { MetaLeaderboard as MetaLeaderboardData } from "@/types";
 import type { ResultRow } from "@/lib/duckdbQueries";
 import { getMetaLeaderboardData, listResults } from "@/lib/duckdbQueries";
-import { BENCHMARK_LABELS, humanizeBenchmark, fmtScoreCompact, fmtScoreExact, fmtGeomean, errMsg } from "@/utils";
+import {
+  BENCHMARK_LABELS,
+  humanizeBenchmark,
+  fmtScoreCompact,
+  fmtScoreExact,
+  fmtGeomean,
+  errMsg,
+} from "@/utils";
 import { canonicalBenchmarkSlug, formatBenchmarkLabel } from "@/lib/displayLabels";
-import { MetaLeaderboardSkeleton, SkeletonBlock } from "@/components/LoadingSpinner";
+import { SkeletonBlock } from "@/components/LoadingSpinner";
 import { ErrorMessage } from "@/components/ErrorMessage";
-import { MetaLeaderboard } from "@/components/MetaLeaderboard";
+import { PageHeader } from "@/components/PageHeader";
 import { TableScrollHint } from "@/components/TableScrollHint";
-import type { MetaLeaderboardMode } from "@/components/MetaLeaderboard";
 import { ProvenanceLegend } from "@/components/ProvenanceLegend";
-import {
-  ALL_FACET_KEYS,
-  facetsToWhereClause,
-  useFacetState,
-  type ExplorerFacetKey,
-  type FacetState,
-} from "@/lib/facetModel";
-import {
-  appendFacetParams,
-  hasActiveFacets,
-  matchesFacetRow,
-  singleFacetValue,
-  toggleFacetValue,
-  toDateWindowFacet,
-} from "@/lib/facetMatching";
-import { formatCount, formatCountWithVerb } from "@/lib/copyFormatters";
-import { normalizedCostLabel, normalizedCostValue } from "@/lib/costDisplay";
-import { formatFacetDisplayValue } from "@/lib/facetDisplay";
-import { stringSerde, useUrlState } from "@/lib/useUrlState";
-import { useDocumentTitle } from "@/lib/useDocumentTitle";
-import { usePickingState } from "@/lib/pickingState";
 import { RunDateChip } from "@/components/RunAge";
-import {
-  EXPLORER_PERFORMANCE_MARKS,
-  EXPLORER_PERFORMANCE_MEASURES,
-  markExplorerPerformance,
-  measureExplorerPerformance,
-} from "@/lib/performanceMarks";
+import { formatCount } from "@/lib/copyFormatters";
+import { normalizedCostLabel, normalizedCostValue } from "@/lib/costDisplay";
+import { useDocumentTitle } from "@/lib/useDocumentTitle";
 
 const SUPPORTED_BENCHMARK_COUNT = new Set(Object.values(BENCHMARK_LABELS)).size;
-const UNLABELLED_TUNING_VALUE = "untuned";
+const RECENT_RESULT_COUNT = 5;
 
-export const HOME_SHELL_GEOMETRY_CLASSES = {
-  heroSurface: "surface-hero border-b border-[var(--bb-border-default)]",
-  heroWrapper: "mx-auto max-w-7xl px-4 py-4 sm:px-6 sm:py-7 lg:px-8 lg:py-4",
-  heroIntro: "max-w-4xl",
-  headline: "text-2xl font-bold sm:text-4xl",
-  subtitle: "mt-2 max-w-3xl text-sm text-[var(--bb-fg-muted)] sm:mt-3 sm:text-lg",
-  activeSummary:
-    "mt-4 rounded-lg border border-[var(--bb-border-default)] bg-[var(--bb-bg-panel)] p-2 sm:mt-5 sm:p-3",
-  activeSummaryItems: "flex flex-wrap gap-2",
-  activeSummaryChip:
-    "inline-flex min-w-0 max-w-full items-center gap-1.5 rounded-full border border-[var(--bb-border-subtle)] bg-[var(--bb-bg-primary)] px-2.5 py-1 text-xs",
-  activeSummaryLabel: "shrink-0 font-semibold uppercase tracking-wide text-[var(--bb-fg-muted)]",
-  activeSummaryValue: "truncate font-medium text-[var(--bb-fg-primary)]",
-  activeFacetList: "mt-3 grid gap-2 sm:flex sm:flex-wrap",
-  activeFacetChip:
-    "block min-w-0 truncate rounded-full bg-[var(--bb-bg-elevated)] px-3 py-1.5 text-xs font-medium text-[var(--bb-fg-muted)]",
-  rankingSelector:
-    "mt-3 rounded-lg border border-[var(--bb-border-default)] bg-[var(--bb-bg-panel)] p-2 sm:mt-5 sm:p-3",
-  rankingGrid: "grid grid-cols-2 gap-2 sm:gap-3 md:grid-cols-3 xl:grid-cols-5",
-  scopeDetails: "mt-2 border-t border-[var(--bb-border-default)] pt-2",
-  scopeSummary:
-    "cursor-pointer text-xs font-medium text-[var(--bb-fg-muted)] hover:text-[var(--bb-fg-primary)]",
-  advancedDetails: "mt-3 border-t border-[var(--bb-border-default)] pt-3",
-  advancedSummary:
-    "cursor-pointer text-sm font-medium text-[var(--bb-fg-muted)] hover:text-[var(--bb-fg-primary)]",
-  dataSurface: "surface-app mx-auto max-w-7xl px-4 py-4 sm:px-6 sm:py-8 lg:px-8",
-} as const;
-
-interface TuningModeSummary {
-  labelledOptions: string[];
-  unlabelledCount: number;
-  labelledCount: number;
-  totalCount: number;
-}
-
+/**
+ * The corpus overview: what BenchBox holds and what arrived most recently.
+ *
+ * The ranking table and its filters live on the compare route, which is where
+ * a reader who wants to rank runs against each other is already headed. This
+ * page answers the prior question - what is in the corpus at all - without
+ * putting a filter between the reader and the answer.
+ */
 export function Home(_: RoutableProps) {
   const recentResultsScrollerRef = useRef<HTMLDivElement>(null);
-  useDocumentTitle("Results · BenchBox");
+  useDocumentTitle("Overview · BenchBox");
   const [results, setResults] = useState<ResultRow[] | null>(null);
-  const [platformVersionDomainResults, setPlatformVersionDomainResults] = useState<ResultRow[] | null>(null);
   const [metaLeaderboard, setMetaLeaderboard] = useState<MetaLeaderboardData | null>(null);
-  const [metaLeaderboardLoaded, setMetaLeaderboardLoaded] = useState(false);
-  const retriedEmptyResults = useRef(false);
-  const [emptyResultsRetryFinished, setEmptyResultsRetryFinished] = useState(false);
   const [error, setError] = useState<string | null>(null);
-  const { facets, where: facetWhere, setFacet, resetFacets } = useFacetState();
-  const [modeRaw, setModeRaw] = useUrlState<string>("mode", "speedup", stringSerde);
-  const benchmarkFilters = facets.benchmark;
-  const scaleFilters = facets.scale_factor;
-  const phaseFilter = singleFacetValue(facets.phase, "all");
-  const tuningFilter = singleFacetValue(facets.tuning_mode, "all");
-  const trustFilter = singleFacetValue(facets.trust_tier, "all");
-  const dateWindow = facets.date_window;
-  const platformVersionDomainWhere = useMemo(
-    () => facetsToWhereClause({ ...facets, platform_version: [] }),
-    [
-      facets.arch,
-      facets.benchmark,
-      facets.cloud_provider,
-      facets.cloud_region,
-      facets.cost_status,
-      facets.cpu_family,
-      facets.date_window,
-      facets.deployment_class,
-      facets.execution_mode,
-      facets.instance_or_warehouse,
-      facets.phase,
-      facets.platform,
-      facets.scale_factor,
-      facets.storage_format,
-      facets.trust_tier,
-      facets.tuning_mode,
-      facets.validation_status,
-    ],
-  );
 
   useEffect(() => {
     let cancelled = false;
-    listResults(facetWhere)
+    // The overview reports the whole corpus, so it reads it unfiltered rather
+    // than through the facet state the ranking table maintains.
+    listResults()
       .then((rows) => {
         if (!cancelled) setResults(rows);
       })
@@ -137,223 +54,27 @@ export function Home(_: RoutableProps) {
     return () => {
       cancelled = true;
     };
-  }, [facetWhere]);
-
-  useEffect(() => {
-    if (facets.platform_version.length === 0) {
-      setPlatformVersionDomainResults(null);
-      return;
-    }
-    let cancelled = false;
-    listResults(platformVersionDomainWhere)
-      .then((rows) => {
-        if (!cancelled) setPlatformVersionDomainResults(rows);
-      })
-      .catch(() => {
-        if (!cancelled) setPlatformVersionDomainResults(null);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [facets.platform_version, platformVersionDomainWhere]);
+  }, []);
 
   useEffect(() => {
     let cancelled = false;
-    markExplorerPerformance(EXPLORER_PERFORMANCE_MARKS.HOME_LEADERBOARD_DATA_START, { once: true });
     getMetaLeaderboardData()
       .then((data) => {
-        if (!cancelled) {
-          markExplorerPerformance(EXPLORER_PERFORMANCE_MARKS.HOME_LEADERBOARD_DATA_READY, {
-            once: true,
-            detail: { cohortCount: data?.cohorts.length ?? 0 },
-          });
-          measureExplorerPerformance(
-            EXPLORER_PERFORMANCE_MEASURES.HOME_LEADERBOARD_DATA,
-            EXPLORER_PERFORMANCE_MARKS.HOME_LEADERBOARD_DATA_START,
-            EXPLORER_PERFORMANCE_MARKS.HOME_LEADERBOARD_DATA_READY,
-            { once: true },
-          );
-          setMetaLeaderboard(data);
-          setMetaLeaderboardLoaded(true);
-        }
+        if (!cancelled) setMetaLeaderboard(data);
       })
       .catch(() => {
-        if (!cancelled) setMetaLeaderboardLoaded(true);
+        // The rankings only supply a count here. Losing them leaves the rest of
+        // the overview usable, so this failure is not surfaced as a page error.
+        if (!cancelled) setMetaLeaderboard(null);
       });
     return () => {
       cancelled = true;
     };
   }, []);
 
-  // Cold-load mitigation (N5 in pass-2 review): unfiltered listResults() can briefly
-  // resolve to [] on a cold DuckDB-WASM attach while the meta-leaderboard
-  // succeeds, which would otherwise paint "0 Results / 0 Benchmarks /
-  // 0 Platforms" before the real corpus arrives. Retrying once is a
-  // symptom-mitigation, not a root-cause fix; we have not pinned which of
-  // (a) WASM cold-cache partial read, (b) Preact effect-cleanup race, or
-  // (c) shared initPromise in db.ts is the trigger. The diagnostic log
-  // makes future investigation possible by surfacing the exact mismatch
-  // in browser DevTools without needing to re-run the audit setup.
-  const hasInconsistentEmptySnapshot =
-    facetWhere.sql === "" && results !== null && results.length === 0 && metaLeaderboard !== null;
-  const hasPersistentInconsistentEmptySnapshot = hasInconsistentEmptySnapshot && emptyResultsRetryFinished;
-
-  useEffect(() => {
-    if (!hasInconsistentEmptySnapshot || retriedEmptyResults.current) return;
-    let cancelled = false;
-    retriedEmptyResults.current = true;
-    if (typeof console !== "undefined") {
-      console.warn(
-        "[Home] Inconsistent cold-load snapshot detected: listResults()=[] but meta-leaderboard has %d cohort(s). Retrying once.",
-        metaLeaderboard?.cohorts?.length ?? 0,
-      );
-    }
-    listResults(facetWhere)
-      .then((rows) => {
-        if (!cancelled) {
-          setResults(rows);
-          setEmptyResultsRetryFinished(true);
-        }
-      })
-      .catch((err: unknown) => {
-        if (!cancelled) {
-          setEmptyResultsRetryFinished(true);
-          setError(errMsg(err));
-        }
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [facetWhere, hasInconsistentEmptySnapshot]);
-
-  const resultById = useMemo(
-    () => new Map((results ?? []).map((result) => [result.result_id, result])),
-    [results],
-  );
-
-  const cohortPlatformIndex = useMemo(() => {
-    const index = new Map<string, Map<string, MetaCohortPlatform[]>>();
-    if (!metaLeaderboard) return index;
-    for (const cohort of metaLeaderboard.cohorts) {
-      const inner = new Map<string, MetaCohortPlatform[]>();
-      for (const row of cohort.platforms ?? []) {
-        const rows = inner.get(row.platform_id) ?? [];
-        rows.push(row);
-        inner.set(row.platform_id, rows);
-      }
-      index.set(cohort.key, inner);
-    }
-    return index;
-  }, [metaLeaderboard]);
-
-  const filteredMetaLeaderboard = useMemo(() => {
-    if (!metaLeaderboard) return null;
-
-    const visibleCohorts = metaLeaderboard.cohorts.map((cohort) => ({
-      ...cohort,
-      platforms: (cohort.platforms ?? []).filter((row) => {
-        const result = resultById.get(row.result_id);
-        return result !== undefined && matchesFacetRow(result, facets);
-      }),
-    })).filter((cohort) => {
-      if (
-        benchmarkFilters.length > 0 &&
-        !benchmarkFilters.map(canonicalBenchmarkSlug).includes(canonicalBenchmarkSlug(cohort.benchmark))
-      ) return false;
-      if (scaleFilters.length > 0 && !scaleFilters.includes(String(cohort.scale_factor))) return false;
-      if (phaseFilter !== "all" && cohort.phase !== phaseFilter) return false;
-
-      return (cohort.platforms ?? []).length > 0;
-    });
-
-    const visibleEvidencePlatformIds = new Set(
-      visibleCohorts.flatMap((cohort) => (cohort.platforms ?? []).map((row) => row.platform_id)),
-    );
-    const platforms = metaLeaderboard.platforms
-      .map((platform) => {
-        const ranks = Object.fromEntries(
-          visibleCohorts.flatMap((cohort) => {
-            const rows = cohortPlatformIndex.get(cohort.key)?.get(platform.platform_id) ?? [];
-            const rankedRows = rows
-              .filter((row) => {
-              const result = resultById.get(row.result_id);
-                return row.rank !== null && result !== undefined && matchesFacetRow(result, facets);
-              })
-              .sort((a, b) => (a.rank ?? Number.POSITIVE_INFINITY) - (b.rank ?? Number.POSITIVE_INFINITY));
-            const row = rankedRows[0];
-            if (!row || row.rank === null) return [];
-            return [[cohort.key, {
-              rank: row.rank,
-              total: cohort.cohort_ranked_count,
-              metric_value: row.metric_value,
-              speedup_vs_best: row.speedup_vs_best,
-            } satisfies MetaRank] as const];
-          }),
-        ) as Record<string, MetaRank>;
-
-        const rankValues = Object.values(ranks).map((rank) => rank.rank);
-        const avgRank = rankValues.length > 0
-          ? Math.round((rankValues.reduce((sum, rank) => sum + rank, 0) / rankValues.length) * 10) / 10
-          : null;
-
-        return {
-          ...platform,
-          ranks,
-          avg_rank: avgRank,
-          n_cohorts: rankValues.length,
-        } satisfies MetaPlatform;
-      })
-      .filter((platform) => platform.n_cohorts > 0 || visibleEvidencePlatformIds.has(platform.platform_id))
-      .sort((a, b) => {
-        if (a.avg_rank === null && b.avg_rank === null) return 0;
-        if (a.avg_rank === null) return 1;
-        if (b.avg_rank === null) return -1;
-        return a.avg_rank - b.avg_rank;
-      });
-
-    return {
-      ...metaLeaderboard,
-      cohorts: visibleCohorts,
-      platforms,
-    };
-  }, [
-    benchmarkFilters,
-    cohortPlatformIndex,
-    facets,
-    metaLeaderboard,
-    phaseFilter,
-    resultById,
-    scaleFilters,
-  ]);
-
   if (error) return <ErrorMessage title="Could not load results" message={error} />;
-  if (hasPersistentInconsistentEmptySnapshot) {
-    return (
-      <ErrorMessage
-        title="Could not load all results"
-        message="The rankings loaded, but the run list did not. Reload the page to try again."
-      />
-    );
-  }
-  if (!results || !metaLeaderboardLoaded || hasInconsistentEmptySnapshot) {
-    return (
-      <HomeLoadingSkeleton
-        benchmark={summarizeSelection(benchmarkFilters, "All benchmarks", humanizeBenchmark)}
-        scaleFactor={summarizeSelection(scaleFilters, "All scales", (value) => `SF ${value}`)}
-        phase={phaseFilter === "all" ? "All phases" : phaseFilter}
-        activeFacets={summarizeActiveFacets(facets, new Map())}
-        showActiveSummary={
-          benchmarkFilters.length > 0 ||
-          scaleFilters.length > 0 ||
-          phaseFilter !== "all" ||
-          hasActiveFacets(facets, RENDERED_FACET_KEYS)
-        }
-      />
-    );
-  }
+  if (!results) return <OverviewSkeleton />;
 
-  const mode: MetaLeaderboardMode =
-    modeRaw === "times" || modeRaw === "ranks" ? modeRaw : "speedup";
   const benchmarks = [...new Set(results.map((result) => canonicalBenchmarkSlug(result.benchmark)))].sort();
   const platformIdToName = new Map(
     (metaLeaderboard?.platforms ?? []).map((platform) => [platform.platform_id, platform.platform]),
@@ -362,948 +83,134 @@ export function Home(_: RoutableProps) {
     platformIdToName.set(result.platform_id, result.platform);
   }
   const platformIds = [...new Set(results.map((result) => result.platform_id))].sort();
-  const activeFacetSummaries = summarizeActiveFacets(facets, platformIdToName);
-  const benchmarkOptions = metaLeaderboard
-    ? [...new Set(metaLeaderboard.cohorts.map((cohort) => cohort.benchmark))].sort()
-    : [];
-  const scaleOptions = metaLeaderboard
-    ? [...new Set(metaLeaderboard.cohorts.map((cohort) => String(cohort.scale_factor)))].sort(
-        (a, b) => Number(a) - Number(b),
-      )
-    : [];
-  const phaseOptions = metaLeaderboard
-    ? [...new Set(metaLeaderboard.cohorts.map((cohort) => cohort.phase))].sort()
-    : [];
-  const trustOptions = ["all", ...[...new Set(results.map((result) => result.trust_label))].sort()];
-  const platformVersionOptions = (() => {
-    const versions = new Set<string>();
-    for (const r of platformVersionDomainResults ?? results) {
-      if (r.platform_version) {
-        versions.add(r.platform_version);
-      }
-    }
-    return [...versions].sort((a, b) => a.localeCompare(b, undefined, { numeric: true }));
-  })();
-  const platformVersionFilters = facets.platform_version;
   const recent = [...results]
     .sort((a, b) => b.run_date.localeCompare(a.run_date))
-    .slice(0, 5);
+    .slice(0, RECENT_RESULT_COUNT);
   const showRecentCost = recent.some((result) => normalizedCostValue(result) !== null);
   const leaderboardCohortCount = metaLeaderboard?.cohorts.length ?? 0;
-  const fullLeaderboardPlatformCount = metaLeaderboard?.platforms.length ?? 0;
-  const rankedLeaderboardPlatformCount =
+  const rankedPlatformCount =
     metaLeaderboard?.platforms.filter((platform) => platform.n_cohorts > 0).length ?? 0;
-  const visibleLeaderboardCohortCount = filteredMetaLeaderboard?.cohorts.length ?? 0;
-  const visibleLeaderboardPlatformCount = filteredMetaLeaderboard?.platforms.length ?? 0;
-  const leaderboardBenchmarkSet = new Set(benchmarkOptions);
-  const publicBenchmarksOutsideLeaderboard = benchmarks.filter((benchmark) => !leaderboardBenchmarkSet.has(canonicalBenchmarkSlug(benchmark)));
-  const leaderboardEvidencePlatformIds = new Set(
-    (metaLeaderboard?.cohorts ?? []).flatMap((cohort) =>
-      (cohort.platforms ?? []).map((platform) => platform.platform_id),
-    ),
-  );
-  const publicPlatformIdsOutsideLeaderboard = platformIds.filter((platformId) => !leaderboardEvidencePlatformIds.has(platformId));
-  const leaderboardEvidencePlatformCount = leaderboardEvidencePlatformIds.size || fullLeaderboardPlatformCount;
-  const visibleRankedLeaderboardPlatformCount =
-    filteredMetaLeaderboard?.platforms.filter((platform) => platform.n_cohorts > 0).length ?? 0;
-  const tuningSummary = summarizeTuningModes(results);
-  // An active tuning filter means `results` was already loaded with that
-  // predicate applied, so the summary necessarily collapses to one bucket.
-  // Dropping the selector then would strand the user on the filtered view with
-  // no control to return to "All tuning labels", so keep it whenever a filter
-  // is active regardless of the narrowed cardinality.
-  const tuningOptions = tuningSummaryBucketCount(tuningSummary) > 1 || tuningFilter !== "all"
-    ? [
-        "all",
-        ...(tuningSummary.unlabelledCount > 0 ? [UNLABELLED_TUNING_VALUE] : []),
-        ...tuningSummary.labelledOptions,
-      ]
-    : [];
-
-  function buildCohortHref(cohort: MetaCohort): string {
-    const params = new URLSearchParams();
-    params.set("sf", String(cohort.scale_factor));
-    params.set("phase", cohort.phase);
-    appendFacetParams(params, facets, new Set(["benchmark", "scale_factor", "phase"]));
-    const query = params.toString();
-    return `/results/${cohort.benchmark}/${query ? `?${query}` : ""}`;
-  }
-
-  function buildPlatformHref(platformId: string): string {
-    const params = new URLSearchParams();
-    appendFacetParams(params, facets, new Set(["platform"]));
-    const query = params.toString();
-    return `/results/p/${platformId}/${query ? `?${query}` : ""}`;
-  }
-
-  const leaderboardScopeSummaryProps = {
-    publicBenchmarkCount: benchmarks.length,
-    leaderboardBenchmarkCount: benchmarkOptions.length,
-    publicBenchmarksOutsideLeaderboard,
-    publicPlatformCount: platformIds.length,
-    leaderboardEvidencePlatformCount,
-    fullLeaderboardPlatformCount,
-    rankedLeaderboardPlatformCount,
-    visibleLeaderboardPlatformCount,
-    visibleRankedLeaderboardPlatformCount,
-    publicPlatformIdsOutsideLeaderboard: publicPlatformIdsOutsideLeaderboard.length,
-    tuningSummary,
-  };
 
   return (
-    <div>
-      <section
-        class={HOME_SHELL_GEOMETRY_CLASSES.heroSurface}
-        data-testid="home-hero-filter-band"
-        data-surface="hero"
-      >
-        {/* lg:py-4 trims hero vertical cost at desktop and wide only. Those two
-            viewports budget the fold at 900px, and the taller hero pushed the
-            second leaderboard row past it (see responsive.spec.ts). Tablet and
-            mobile keep the roomier padding; their fold budget is 1200px. */}
-        <div class={HOME_SHELL_GEOMETRY_CLASSES.heroWrapper} data-testid="home-hero-wrapper">
-          <div class={HOME_SHELL_GEOMETRY_CLASSES.heroIntro} data-testid="home-hero-intro">
-            <h1 class={HOME_SHELL_GEOMETRY_CLASSES.headline}>Compare benchmark results</h1>
-            <p class={HOME_SHELL_GEOMETRY_CLASSES.subtitle}>
-              See how published platform runs compare across BenchBox rankings. Open any result to inspect its evidence.
-            </p>
-          </div>
+    <div class="mx-auto max-w-7xl px-4 py-4 sm:px-6 sm:py-8 lg:px-8" data-testid="overview-surface">
+      <PageHeader
+        eyebrow="Results"
+        title="Overview"
+        subtitle="What the public corpus holds today, and what arrived most recently. Rank runs against each other on the compare page."
+        actions={
+          <a href="/results/compare" class="btn btn-primary no-underline" data-testid="overview-compare-cta">
+            Compare benchmark results
+          </a>
+        }
+      />
 
-          {filteredMetaLeaderboard &&
-            (benchmarkFilters.length > 0 ||
-              scaleFilters.length > 0 ||
-              phaseFilter !== "all" ||
-              activeFacetSummaries.length > 0) && (
-              <ActiveLeaderboardSummary
-                benchmark={summarizeSelection(benchmarkFilters, "All benchmarks", humanizeBenchmark)}
-                scaleFactor={summarizeSelection(scaleFilters, "All scales", (value) => `SF ${value}`)}
-                phase={phaseFilter === "all" ? "All phases" : phaseFilter}
-                activeFacets={activeFacetSummaries}
-              />
-            )}
+      <FlywheelStrip />
 
-          {/* Ranking selector renders above the matrix so users can change
-              benchmark/scale/phase context without scrolling past the
-              matrix to find the controls. */}
-          {filteredMetaLeaderboard && (
-            <section
-              aria-label="Leaderboard ranking selector"
-              class={HOME_SHELL_GEOMETRY_CLASSES.rankingSelector}
-            >
-              <div class={HOME_SHELL_GEOMETRY_CLASSES.rankingGrid} data-testid="home-ranking-selector-grid">
-                <MultiSelectFilter
-                  label="Benchmark"
-                  allLabel="All benchmarks"
-                  options={benchmarkOptions}
-                  current={benchmarkFilters}
-                  onSelect={(value) =>
-                    setFacet(
-                      "benchmark",
-                      value === "all" ? [] : toggleFacetValue(benchmarkFilters, value),
-                    )
-                  }
-                  format={(value) => formatBenchmarkLabel(value)}
-                  description={benchmarkFilterDescription(publicBenchmarksOutsideLeaderboard)}
-                />
-                <MultiSelectFilter
-                  label="Scale"
-                  allLabel="All scales"
-                  options={scaleOptions}
-                  current={scaleFilters}
-                  onSelect={(value) =>
-                    setFacet(
-                      "scale_factor",
-                      value === "all" ? [] : toggleFacetValue(scaleFilters, value),
-                    )
-                  }
-                  format={(value) => `SF ${value}`}
-                />
-                <SelectFilter
-                  label="Phase"
-                  options={["all", ...phaseOptions]}
-                  current={phaseFilter}
-                  onSelect={(value) => setFacet("phase", value === "all" ? [] : [value])}
-                  format={(value) => (value === "all" ? "All phases" : value)}
-                />
-                {platformVersionOptions.length > 0 || platformVersionFilters.length > 0 ? (
-                  <MultiSelectFilter
-                    label="Platform version"
-                    allLabel="All versions"
-                    options={platformVersionOptions}
-                    current={platformVersionFilters}
-                    onSelect={(value) =>
-                      setFacet(
-                        "platform_version",
-                        value === "all" ? [] : toggleFacetValue(platformVersionFilters, value),
-                      )
-                    }
-                    format={(value) => value}
-                  />
-                ) : (
-                  <div aria-hidden="true" data-testid="home-engine-version-unavailable" />
-                )}
-                <CoverageSummary />
-              </div>
-
-              <details
-                class={HOME_SHELL_GEOMETRY_CLASSES.scopeDetails}
-                data-testid="leaderboard-scope-summary-mobile"
-              >
-                <summary class={HOME_SHELL_GEOMETRY_CLASSES.scopeSummary}>
-                  What counts as a ranked result?
-                </summary>
-                <LeaderboardScopeSummary {...leaderboardScopeSummaryProps} />
-              </details>
-
-              <details
-                class={HOME_SHELL_GEOMETRY_CLASSES.advancedDetails}
-                data-testid="leaderboard-advanced-filters"
-              >
-                <summary class={HOME_SHELL_GEOMETRY_CLASSES.advancedSummary}>
-                  Advanced filters
-                </summary>
-                <div class="mt-3 grid gap-3 md:grid-cols-3">
-                  {tuningOptions.length > 0 ? (
-                    <SingleFilterGroup
-                      label="Tuning"
-                      options={tuningOptions}
-                      current={tuningFilter}
-                      onSelect={(value) => setFacet("tuning_mode", value === "all" ? [] : [value])}
-                      format={(value) => formatTuningModeOption(value)}
-                      optionTitle={(value) => tuningModeOptionTitle(value, tuningSummary)}
-                      description={tuningFilterDescription(tuningSummary)}
-                    />
-                  ) : (
-                    <div data-testid="tuning-filter-unavailable">
-                      <div class="mb-2 text-xs font-semibold uppercase tracking-wide text-[var(--bb-fg-muted)]">Tuning</div>
-                      <p class="text-[11px] leading-snug text-[var(--bb-fg-muted)]">
-                        Tuning filter unavailable: this view has only one tuning metadata state. {tuningFilterDescription(tuningSummary)}
-                      </p>
-                    </div>
-                  )}
-                  <SingleFilterGroup
-                    label="Trust"
-                    options={trustOptions}
-                    current={trustFilter}
-                    onSelect={(value) => setFacet("trust_tier", value === "all" ? [] : [value])}
-                    format={(value) => (value === "all" ? "All trust tiers" : value)}
-                  />
-                  <SingleFilterGroup
-                    label="Date window"
-                    options={["all", "30d", "90d", "365d"]}
-                    current={dateWindow}
-                    onSelect={(value) => setFacet("date_window", toDateWindowFacet(value))}
-                    format={(value) => (value === "all" ? "All time" : `Last ${value}`)}
-                  />
-                </div>
-              </details>
-            </section>
-          )}
-        </div>
+      <section aria-label="Corpus summary" class="mb-8 grid grid-cols-2 gap-4 text-center sm:mb-12 lg:grid-cols-4">
+        <StatCard
+          value={SUPPORTED_BENCHMARK_COUNT}
+          label={{ singular: "supported benchmark", plural: "supported benchmarks" }}
+          detail={`${benchmarks.length} with public results`}
+        />
+        <StatCard
+          value={results.length}
+          label={{ singular: "published run", plural: "published runs" }}
+          detail="reviewed by BenchBox maintainers"
+        />
+        <StatCard
+          value={platformIds.length}
+          label={{
+            singular: "platform with public results",
+            plural: "platforms with public results",
+          }}
+          detail="published platform IDs"
+        />
+        <StatCard
+          value={leaderboardCohortCount}
+          label={{ singular: "ranking", plural: "rankings" }}
+          detail={`${formatCount(rankedPlatformCount, "ranked platform")}`}
+        />
       </section>
 
-      <div
-        class={HOME_SHELL_GEOMETRY_CLASSES.dataSurface}
-        data-testid="home-data-surface"
-        data-surface="app"
-      >
-        {filteredMetaLeaderboard && (
-          <MetaLeaderboard
-            data={filteredMetaLeaderboard}
-            mode={mode}
-            onModeChange={(value) => setModeRaw(value)}
-            cohortHref={buildCohortHref}
-            platformHref={buildPlatformHref}
-            resultMetadataById={resultById}
+      <section class="mb-8 sm:mb-12">
+        <h2 class="mb-4 text-xl font-semibold text-[var(--bb-data-fg-primary)]">Recent results</h2>
+        <div class="overflow-hidden rounded-lg border border-[var(--bb-data-border)] bg-[var(--bb-surface-data)] shadow-sm">
+          <TableScrollHint
+            scrollerRef={recentResultsScrollerRef}
+            testId="recent-results-scroll-hint"
+            label="Scroll table for row actions →"
+            wrapperClassName="flex justify-end"
+            className="m-2"
           />
-        )}
-
-        {filteredMetaLeaderboard && filteredMetaLeaderboard.cohorts.length === 0 && (
-          <CoverageEmptyState
-            activeFacets={activeFacetSummaries}
-            canClearScale={scaleFilters.length > 0}
-            canClearPlatform={facets.platform.length > 0}
-            onClearScale={() => setFacet("scale_factor", [])}
-            onClearPlatform={() => setFacet("platform", [])}
-            onReset={resetFacets}
-          />
-        )}
-
-        {filteredMetaLeaderboard && <FlywheelStrip />}
-
-        <section aria-label="Corpus summary" class="mb-8 grid grid-cols-2 gap-4 text-center sm:mb-12 lg:grid-cols-4">
-          <StatCard
-            value={SUPPORTED_BENCHMARK_COUNT}
-            label={{ singular: "supported benchmark", plural: "supported benchmarks" }}
-            detail={`${benchmarks.length} with public results`}
-          />
-          <StatCard
-            value={results.length}
-            label={{ singular: "published run", plural: "published runs" }}
-            detail="reviewed by BenchBox maintainers"
-          />
-          <StatCard
-            value={platformIds.length}
-            label={{
-              singular: "platform with public results",
-              plural: "platforms with public results",
-            }}
-            detail="published platform IDs"
-          />
-          <StatCard
-            value={leaderboardCohortCount}
-            label={{ singular: "leaderboard ranking", plural: "leaderboard rankings" }}
-            detail={`${visibleLeaderboardCohortCount} visible; ${visibleLeaderboardPlatformCount}/${leaderboardEvidencePlatformCount} ranked-scope platforms`}
-          />
-        </section>
-
-        <section class="mb-8 sm:mb-12">
-          <h2 class="mb-4 text-xl font-semibold text-[var(--bb-data-fg-primary)]">Recent results</h2>
-          <div class="overflow-hidden rounded-lg border border-[var(--bb-data-border)] bg-[var(--bb-surface-data)] shadow-sm">
-            <TableScrollHint
-              scrollerRef={recentResultsScrollerRef}
-              testId="recent-results-scroll-hint"
-              label="Scroll table for row actions →"
-              wrapperClassName="flex justify-end"
-              className="m-2"
-            />
-            <div ref={recentResultsScrollerRef} class="overflow-x-auto" data-testid="recent-results-scroll-container">
-              <table class="min-w-full w-max divide-y divide-[var(--bb-data-border)]">
-                <thead class="bg-[var(--bb-surface-data-muted)]">
-                  <tr>
-                    <th class="table-th">Benchmark</th>
-                    <th class="table-th">Platform</th>
-                    <th class="table-th">Scale</th>
-                    <th class="table-th">Date</th>
-                    <th class="table-th">Power score</th>
+          <div ref={recentResultsScrollerRef} class="overflow-x-auto" data-testid="recent-results-scroll-container">
+            <table class="min-w-full w-max divide-y divide-[var(--bb-data-border)]">
+              <thead class="bg-[var(--bb-surface-data-muted)]">
+                <tr>
+                  <th class="table-th">Benchmark</th>
+                  <th class="table-th">Platform</th>
+                  <th class="table-th">Scale</th>
+                  <th class="table-th">Date</th>
+                  <th class="table-th">Power score</th>
+                  <th
+                    class="table-th"
+                    title="Geometric mean of per-query execution times (measurement runs only). Lower is faster."
+                  >
+                    Geomean latency
+                  </th>
+                  {showRecentCost && (
                     <th
                       class="table-th"
-                      title="Geometric mean of per-query execution times (measurement runs only). Lower is faster."
+                      title="BenchBox estimated this cloud cost from the recorded deployment details. Local runs and runs without cost details are not included."
                     >
-                      Geomean latency
+                      Normalized cost
                     </th>
-                    {showRecentCost && (
-                      <th
-                        class="table-th"
-                        title="BenchBox estimated this cloud cost from the recorded deployment details. Local runs and runs without cost details are not included."
-                      >
-                        Normalized cost
-                      </th>
-                    )}
-                    <th class="table-th sticky right-0 z-10 bg-[var(--bb-surface-data-muted)]" />
-                  </tr>
-                </thead>
-                <tbody class="divide-y divide-[var(--bb-data-border)] bg-[var(--bb-surface-data)]">
-                  {recent.map((result) => (
-                    <RecentRow key={result.result_id} entry={result} showCost={showRecentCost} />
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </div>
-        </section>
-
-        <div class="grid grid-cols-1 gap-8 sm:grid-cols-2">
-          <BrowseSection
-            title="Browse public benchmark results"
-            description={`${formatCount(benchmarks.length, "public benchmark set")}. Leaderboard filters above include ${formatCount(leaderboardCohortCount, "ranked leaderboard")}.`}
-            items={benchmarks}
-            hrefFn={(benchmark) => `/results/${benchmark}/`}
-            labelFn={formatBenchmarkLabel}
-          />
-          <BrowseSection
-            title="Browse public platform results"
-            description={`${formatCount(platformIds.length, "published platform ID")} in the public corpus, independent of current leaderboard coverage.`}
-            items={platformIds}
-            hrefFn={(platformId) => `/results/p/${platformId}/`}
-            labelFn={(platformId) => platformIdToName.get(platformId) ?? platformId}
-          />
-        </div>
-      </div>
-      <ProvenanceLegend />
-</div>
-  );
-}
-
-function HomeLoadingSkeleton({
-  benchmark,
-  scaleFactor,
-  phase,
-  activeFacets,
-  showActiveSummary,
-}: {
-  benchmark: string;
-  scaleFactor: string;
-  phase: string;
-  activeFacets: ActiveFacetSummary[];
-  showActiveSummary: boolean;
-}) {
-  return (
-    <div>
-      <section
-        class={HOME_SHELL_GEOMETRY_CLASSES.heroSurface}
-        data-testid="home-hero-filter-band"
-        data-surface="hero"
-      >
-        {/* Loaded/skeleton geometry is owned by HOME_SHELL_GEOMETRY_CLASSES. */}
-        <div class={HOME_SHELL_GEOMETRY_CLASSES.heroWrapper} data-testid="home-hero-wrapper">
-          <div class={HOME_SHELL_GEOMETRY_CLASSES.heroIntro} data-testid="home-hero-intro">
-            <h1 class={HOME_SHELL_GEOMETRY_CLASSES.headline}>Compare benchmark results</h1>
-            <p class={HOME_SHELL_GEOMETRY_CLASSES.subtitle}>
-              See how published platform runs compare across BenchBox rankings. Open any result to inspect its evidence.
-            </p>
-          </div>
-
-          {showActiveSummary && (
-            <div
-              aria-hidden="true"
-              class={`${HOME_SHELL_GEOMETRY_CLASSES.activeSummary} sm:hidden`}
-              data-testid="home-loading-active-summary-reserve"
-            >
-              <div class={HOME_SHELL_GEOMETRY_CLASSES.activeSummaryItems}>
-                {[
-                  ["Benchmark", benchmark],
-                  ["Scale", scaleFactor],
-                  ["Phase", phase],
-                ].map(([label, value]) => (
-                  <div key={label} class={`${HOME_SHELL_GEOMETRY_CLASSES.activeSummaryChip} bb-skeleton-dark`}>
-                    <span class={`${HOME_SHELL_GEOMETRY_CLASSES.activeSummaryLabel} invisible`}>{label}</span>
-                    <span class={`${HOME_SHELL_GEOMETRY_CLASSES.activeSummaryValue} invisible`}>{value}</span>
-                  </div>
+                  )}
+                  <th class="table-th sticky right-0 z-10 bg-[var(--bb-surface-data-muted)]" />
+                </tr>
+              </thead>
+              <tbody class="divide-y divide-[var(--bb-data-border)] bg-[var(--bb-surface-data)]">
+                {recent.map((result) => (
+                  <RecentRow key={result.result_id} entry={result} showCost={showRecentCost} />
                 ))}
-              </div>
-              {activeFacets.length > 0 && (
-                <div class={HOME_SHELL_GEOMETRY_CLASSES.activeFacetList}>
-                  {activeFacets.map((facet) => (
-                    <span
-                      key={facet.key}
-                      class={`${HOME_SHELL_GEOMETRY_CLASSES.activeFacetChip} bb-skeleton-dark`}
-                    >
-                      <span class="invisible">{facet.label}: {facet.value}</span>
-                    </span>
-                  ))}
-                </div>
-              )}
-            </div>
-          )}
-
-          <section
-            aria-label="Leaderboard ranking selector"
-            class={HOME_SHELL_GEOMETRY_CLASSES.rankingSelector}
-          >
-            <div class={HOME_SHELL_GEOMETRY_CLASSES.rankingGrid} data-testid="home-ranking-selector-grid">
-              <SkeletonSelect label="Benchmark" />
-              <SkeletonSelect label="Scale" />
-              <SkeletonSelect label="Phase" />
-              <SkeletonSelect label="Platform version" />
-              <CoverageSummary />
-            </div>
-            <div aria-hidden="true" class="sm:hidden">
-              <div
-                class={HOME_SHELL_GEOMETRY_CLASSES.scopeDetails}
-                data-testid="home-loading-scope-details-reserve"
-              >
-                <SkeletonBlock className="h-4 w-40 bb-skeleton-dark" />
-              </div>
-              <div
-                class={HOME_SHELL_GEOMETRY_CLASSES.advancedDetails}
-                data-testid="home-loading-advanced-details-reserve"
-              >
-                <SkeletonBlock className="h-5 w-32 bb-skeleton-dark" />
-              </div>
-            </div>
-          </section>
+              </tbody>
+            </table>
+          </div>
         </div>
       </section>
 
-      <div
-        class={HOME_SHELL_GEOMETRY_CLASSES.dataSurface}
-        data-testid="home-data-surface"
-        data-surface="app"
-      >
-        <MetaLeaderboardSkeleton />
+      <div class="grid grid-cols-1 gap-8 sm:grid-cols-2">
+        <BrowseSection
+          title="Browse public benchmark results"
+          description={`${formatCount(benchmarks.length, "public benchmark set")}, covered by ${formatCount(leaderboardCohortCount, "ranking")}.`}
+          items={benchmarks}
+          hrefFn={(benchmark) => `/results/${benchmark}/`}
+          labelFn={formatBenchmarkLabel}
+        />
+        <BrowseSection
+          title="Browse public platform results"
+          description={`${formatCount(platformIds.length, "published platform ID")} in the public corpus, independent of ranking coverage.`}
+          items={platformIds}
+          hrefFn={(platformId) => `/results/p/${platformId}/`}
+          labelFn={(platformId) => platformIdToName.get(platformId) ?? platformId}
+        />
       </div>
+
+      <ProvenanceLegend />
     </div>
   );
 }
 
-function SkeletonSelect({ label }: { label: string }) {
+function OverviewSkeleton() {
   return (
-    <div>
-      <div class="mb-2 text-xs font-semibold uppercase tracking-wide text-[var(--bb-fg-muted)]">{label}</div>
-      <SkeletonBlock className="h-9 w-full bb-skeleton-dark" />
-    </div>
-  );
-}
-
-interface ActiveFacetSummary {
-  // Widened to ExplorerFacetKey so the engine-version chip, which is a
-  // hardware facet key, can be summarized alongside the core ones.
-  key: ExplorerFacetKey;
-  label: string;
-  value: string;
-}
-
-/**
- * Facets rendered as active chips on this page.
- *
- * Every URL-supported facet is summarized when active. Architecture and CPU
- * family do not need always-visible controls to be honest URL state: a shared
- * link that carries either filter must still show the reader what constrained
- * the results.
- */
-const RENDERED_FACET_KEYS = ALL_FACET_KEYS;
-
-const FACET_LABELS: Record<ExplorerFacetKey, string> = {
-  benchmark: "Benchmark",
-  scale_factor: "Scale factor",
-  phase: "Phase",
-  platform: "Platform",
-  execution_mode: "Execution",
-  tuning_mode: "Tuning",
-  trust_tier: "Trust",
-  validation_status: "Validation",
-  deployment_class: "Deployment",
-  cloud_provider: "Cloud provider",
-  cloud_region: "Cloud region",
-  instance_or_warehouse: "Instance / warehouse",
-  storage_format: "Storage format",
-  cost_status: "Cost status",
-  date_window: "Date window",
-  platform_version: "Platform version",
-  arch: "Architecture",
-  cpu_family: "CPU family",
-};
-
-function ActiveLeaderboardSummary({
-  benchmark,
-  scaleFactor,
-  phase,
-  activeFacets,
-}: {
-  benchmark: string;
-  scaleFactor: string;
-  phase: string;
-  activeFacets: ActiveFacetSummary[];
-}) {
-  const items = [
-    { label: "Benchmark", value: benchmark },
-    { label: "Scale", value: scaleFactor },
-    { label: "Phase", value: phase },
-  ];
-
-  return (
-    <section
-      aria-label="Active leaderboard filters"
-      class={HOME_SHELL_GEOMETRY_CLASSES.activeSummary}
-    >
-      <dl class={HOME_SHELL_GEOMETRY_CLASSES.activeSummaryItems}>
-        {items.map((item) => (
-          <div
-            key={item.label}
-            class={HOME_SHELL_GEOMETRY_CLASSES.activeSummaryChip}
-            title={`${item.label}: ${item.value}`}
-          >
-            <dt class={HOME_SHELL_GEOMETRY_CLASSES.activeSummaryLabel}>{item.label}</dt>
-            <dd class={HOME_SHELL_GEOMETRY_CLASSES.activeSummaryValue}>{item.value}</dd>
-          </div>
+    <div class="mx-auto max-w-7xl px-4 py-4 sm:px-6 sm:py-8 lg:px-8" data-testid="overview-skeleton">
+      <SkeletonBlock className="h-9 w-48" />
+      <SkeletonBlock className="mt-3 h-5 w-full max-w-2xl" />
+      <div class="mt-8 grid grid-cols-2 gap-4 lg:grid-cols-4">
+        {[0, 1, 2, 3].map((slot) => (
+          <SkeletonBlock key={slot} className="h-28" />
         ))}
-      </dl>
-      {activeFacets.length > 0 && (
-        <div aria-label="Active filter chips" class={HOME_SHELL_GEOMETRY_CLASSES.activeFacetList}>
-          {activeFacets.map((facet) => (
-            <span
-              key={facet.key}
-              class={HOME_SHELL_GEOMETRY_CLASSES.activeFacetChip}
-            >
-              {facet.label}: {facet.value}
-            </span>
-          ))}
-        </div>
-      )}
-    </section>
-  );
-}
-
-function CoverageEmptyState({
-  activeFacets,
-  canClearScale,
-  canClearPlatform,
-  onClearScale,
-  onClearPlatform,
-  onReset,
-}: {
-  activeFacets: ActiveFacetSummary[];
-  canClearScale: boolean;
-  canClearPlatform: boolean;
-  onClearScale: () => void;
-  onClearPlatform: () => void;
-  onReset: () => void;
-}) {
-  return (
-    <section
-      aria-labelledby="coverage-empty-title"
-      class="mb-12 rounded-lg border border-dashed border-[var(--bb-data-border-strong)] bg-[var(--bb-surface-data)] p-8 text-[var(--bb-data-fg-muted)]"
-    >
-      <div class="mx-auto max-w-3xl text-center">
-        <h2 id="coverage-empty-title" class="text-lg font-semibold text-[var(--bb-data-fg-primary)]">
-          No leaderboard cells match the current filters
-        </h2>
-        <p class="mt-2 text-sm text-[var(--bb-data-fg-muted)]">
-          The current facet combination removed every published ranking cell.
-        </p>
       </div>
-
-      {activeFacets.length > 0 ? (
-        <dl
-          aria-label="Active filters removing leaderboard cells"
-          class="mx-auto mt-5 grid max-w-3xl gap-2 sm:grid-cols-2 lg:grid-cols-3"
-        >
-          {activeFacets.map((facet) => (
-            <div key={facet.key} class="rounded-md border border-[var(--bb-data-border)] bg-[var(--bb-surface-data-muted)] px-3 py-2 text-left">
-              <dt class="text-[11px] font-semibold uppercase tracking-wide text-[var(--bb-data-fg-subtle)]">{facet.label}</dt>
-              <dd class="mt-1 text-sm font-medium text-[var(--bb-data-fg-primary)]">{facet.value}</dd>
-            </div>
-          ))}
-        </dl>
-      ) : (
-        <p class="mt-5 text-center text-sm text-[var(--bb-data-fg-muted)]">
-          No active facets are applied; the public corpus has no leaderboard rankings for this view.
-        </p>
-      )}
-
-      <div class="mt-5 flex flex-wrap justify-center gap-2">
-        <button
-          type="button"
-          class="rounded-md border border-[var(--bb-data-border-strong)] bg-[var(--bb-surface-data)] px-3 py-2 text-sm font-medium text-[var(--bb-data-fg-primary)] hover:bg-[var(--bb-surface-data-muted)] disabled:cursor-not-allowed disabled:opacity-50"
-          onClick={onClearScale}
-          disabled={!canClearScale}
-        >
-          Clear scale factor
-        </button>
-        <button
-          type="button"
-          class="rounded-md border border-[var(--bb-data-border-strong)] bg-[var(--bb-surface-data)] px-3 py-2 text-sm font-medium text-[var(--bb-data-fg-primary)] hover:bg-[var(--bb-surface-data-muted)] disabled:cursor-not-allowed disabled:opacity-50"
-          onClick={onClearPlatform}
-          disabled={!canClearPlatform}
-        >
-          Clear platform
-        </button>
-        <button
-          type="button"
-          class="btn btn-primary"
-          onClick={onReset}
-        >
-          Reset all
-        </button>
-      </div>
-    </section>
-  );
-}
-
-function LeaderboardScopeSummary({
-  publicBenchmarkCount,
-  leaderboardBenchmarkCount,
-  publicBenchmarksOutsideLeaderboard,
-  publicPlatformCount,
-  leaderboardEvidencePlatformCount,
-  fullLeaderboardPlatformCount,
-  rankedLeaderboardPlatformCount,
-  visibleLeaderboardPlatformCount,
-  visibleRankedLeaderboardPlatformCount,
-  publicPlatformIdsOutsideLeaderboard,
-  tuningSummary,
-}: {
-  publicBenchmarkCount: number;
-  leaderboardBenchmarkCount: number;
-  publicBenchmarksOutsideLeaderboard: string[];
-  publicPlatformCount: number;
-  leaderboardEvidencePlatformCount: number;
-  fullLeaderboardPlatformCount: number;
-  rankedLeaderboardPlatformCount: number;
-  visibleLeaderboardPlatformCount: number;
-  visibleRankedLeaderboardPlatformCount: number;
-  publicPlatformIdsOutsideLeaderboard: number;
-  tuningSummary: TuningModeSummary;
-}) {
-  const hiddenBenchmarkCopy =
-    publicBenchmarksOutsideLeaderboard.length > 0
-      ? `${formatCountWithVerb(publicBenchmarksOutsideLeaderboard.length, "published benchmark", "has", "have")} runs but no ranking: ${publicBenchmarksOutsideLeaderboard.map(formatBenchmarkLabel).join(", ")}.`
-      : "Every benchmark with published runs has a ranking.";
-  const publicOnlyPlatformCopy =
-    publicPlatformIdsOutsideLeaderboard > 0
-      ? `${formatCountWithVerb(publicPlatformIdsOutsideLeaderboard, "published platform", "is", "are")} not represented in the rankings.`
-      : "Every published platform is represented in a ranking.";
-  const visiblePublishedUnrankedPlatformCount = Math.max(
-    0,
-    visibleLeaderboardPlatformCount - visibleRankedLeaderboardPlatformCount,
-  );
-  const fullPublishedUnrankedPlatformCount = Math.max(
-    0,
-    fullLeaderboardPlatformCount - rankedLeaderboardPlatformCount,
-  );
-
-  return (
-    <div class="mt-3 space-y-1 border-t border-[var(--bb-border-default)] pt-3 text-xs text-[var(--bb-fg-muted)]">
-      <p>
-        Ranked results cover {leaderboardBenchmarkCount.toLocaleString()} of{" "}
-        {publicBenchmarkCount.toLocaleString()} published {plural(publicBenchmarkCount, "benchmark", "benchmarks")}.{" "}
-        {hiddenBenchmarkCopy}
-      </p>
-      <p>
-        The current table includes {visibleLeaderboardPlatformCount.toLocaleString()} of {leaderboardEvidencePlatformCount.toLocaleString()} platforms found in the selected rankings: {visibleRankedLeaderboardPlatformCount.toLocaleString()} ranked and {visiblePublishedUnrankedPlatformCount.toLocaleString()} unranked. Across all rankings, {rankedLeaderboardPlatformCount.toLocaleString()} of {fullLeaderboardPlatformCount.toLocaleString()} platforms are ranked and {fullPublishedUnrankedPlatformCount.toLocaleString()} {fullPublishedUnrankedPlatformCount === 1 ? "is" : "are"} unranked. There are {publicPlatformCount.toLocaleString()} published {plural(publicPlatformCount, "platform", "platforms")} in total. {publicOnlyPlatformCopy}
-      </p>
-      <p>
-        Tuning was recorded for {tuningSummary.labelledCount.toLocaleString()} {plural(tuningSummary.labelledCount, "run", "runs")} across {tuningSummary.labelledOptions.length.toLocaleString()} {plural(tuningSummary.labelledOptions.length, "setting", "settings")}. It was not recorded for {tuningSummary.unlabelledCount.toLocaleString()} {plural(tuningSummary.unlabelledCount, "run", "runs")}.
-      </p>
-    </div>
-  );
-}
-
-function summarizeSelection(
-  values: string[],
-  allLabel: string,
-  format: (value: string) => string,
-): string {
-  if (values.length === 0) return allLabel;
-  if (values.length === 1) return format(values[0] ?? "");
-  return `${values.length} selected`;
-}
-
-function summarizeActiveFacets(
-  facets: FacetState,
-  platformIdToName: ReadonlyMap<string, string>,
-): ActiveFacetSummary[] {
-  const summaries: ActiveFacetSummary[] = [];
-  for (const key of RENDERED_FACET_KEYS) {
-    const value = facets[key];
-    if (Array.isArray(value)) {
-      if (value.length === 0) continue;
-      summaries.push({
-        key,
-        label: FACET_LABELS[key],
-        value: value.map((item) => formatFacetValue(key, item, platformIdToName)).join(", "),
-      });
-    } else if (value !== "all") {
-      summaries.push({
-        key,
-        label: FACET_LABELS[key],
-        value: formatFacetValue(key, value, platformIdToName),
-      });
-    }
-  }
-  return summaries;
-}
-
-function formatFacetValue(
-  key: ExplorerFacetKey,
-  value: string,
-  platformIdToName: ReadonlyMap<string, string>,
-): string {
-  if (key === "benchmark") return humanizeBenchmark(value);
-  if (key === "platform") return platformIdToName.get(value) ?? value;
-  if (key === "tuning_mode" && value === UNLABELLED_TUNING_VALUE) return "not recorded";
-  return formatFacetDisplayValue(key, value);
-}
-
-function summarizeTuningModes(results: readonly ResultRow[]): TuningModeSummary {
-  const labelled = new Map<string, number>();
-  let unlabelledCount = 0;
-  for (const result of results) {
-    const value = result.tuning_mode;
-    if (value !== null && value !== undefined && value.trim() !== "") {
-      labelled.set(value, (labelled.get(value) ?? 0) + 1);
-    } else {
-      unlabelledCount += 1;
-    }
-  }
-  const labelledOptions = [...labelled.keys()].sort();
-  const labelledCount = [...labelled.values()].reduce((sum, count) => sum + count, 0);
-  return {
-    labelledOptions,
-    unlabelledCount,
-    labelledCount,
-    totalCount: results.length,
-  };
-}
-
-function benchmarkFilterDescription(publicBenchmarksOutsideLeaderboard: readonly string[]): string {
-  if (publicBenchmarksOutsideLeaderboard.length === 0) {
-    return "Only benchmarks with ranked leaderboards appear here.";
-  }
-  return "Choose a benchmark with ranked results. Open “What counts as a ranked result?” to see other published benchmarks.";
-}
-
-function formatTuningModeOption(value: string): string {
-  if (value === "all") return "All tuning labels";
-  if (value === UNLABELLED_TUNING_VALUE) return "Not Recorded";
-  return formatFacetDisplayValue("tuning_mode", value);
-}
-
-function tuningModeOptionTitle(value: string, summary: TuningModeSummary): string {
-  if (value === "all") return "Include runs with and without recorded tuning details.";
-  if (value === UNLABELLED_TUNING_VALUE) {
-    return `${formatCountWithVerb(summary.unlabelledCount, "public result", "has", "have")} no recorded tuning mode.`;
-  }
-  return `Filter to tuning label: ${formatFacetDisplayValue("tuning_mode", value)}.`;
-}
-
-function tuningFilterDescription(summary: TuningModeSummary): string {
-  if (summary.unlabelledCount === 0) return "All public results in this view have a recorded tuning mode.";
-  if (summary.labelledOptions.length === 0) {
-    return `${formatCountWithVerb(summary.unlabelledCount, "public result", "has", "have")} no recorded tuning mode.`;
-  }
-  return `${summary.unlabelledCount.toLocaleString()} of ${summary.totalCount.toLocaleString()} public ${plural(summary.totalCount, "result", "results")} ${plural(summary.unlabelledCount, "is", "are")} not recorded for tuning.`;
-}
-
-function tuningSummaryBucketCount(summary: TuningModeSummary): number {
-  return summary.labelledOptions.length + (summary.unlabelledCount > 0 ? 1 : 0);
-}
-
-function plural(count: number, singular: string, pluralValue: string): string {
-  return count === 1 ? singular : pluralValue;
-}
-
-type StatCardLabel = string | { singular: string; plural: string };
-
-function resolveStatCardLabel(value: number | string, label: StatCardLabel): string {
-  if (typeof label === "string") return label;
-  return typeof value === "number" && value === 1 ? label.singular : label.plural;
-}
-
-function StatCard({
-  value,
-  label,
-  detail,
-}: {
-  value: number | string;
-  label: StatCardLabel;
-  detail?: string;
-}) {
-  const resolvedLabel = resolveStatCardLabel(value, label);
-  return (
-    <div class="card">
-      <div class="text-3xl font-bold text-[var(--bb-accent-hover)]">{value}</div>
-      <div class="mt-1 text-sm font-medium text-[var(--bb-data-fg-muted)]">{resolvedLabel}</div>
-      {detail && <div class="mt-2 text-xs text-[var(--bb-data-fg-subtle)]">{detail}</div>}
-    </div>
-  );
-}
-
-function MultiSelectFilter({
-  label,
-  allLabel,
-  options,
-  current,
-  onSelect,
-  format,
-  description,
-}: {
-  label: string;
-  allLabel: string;
-  options: string[];
-  current: string[];
-  onSelect: (value: string) => void;
-  format: (value: string) => string;
-  description?: string;
-}) {
-  if (options.length === 0) return null;
-  const value = current.length === 0 ? "all" : current.length === 1 ? current[0] : "__multiple";
-
-  return (
-    <label class="block">
-      <span class="mb-2 block text-xs font-semibold uppercase tracking-wide text-[var(--bb-fg-muted)]">{label}</span>
-      <select
-        aria-label={label}
-        class="h-9 w-full rounded-md border border-[var(--bb-border-default)] bg-[var(--bb-bg-primary)] px-3 text-sm font-medium text-[var(--bb-fg-primary)]"
-        value={value}
-        onChange={(event) => onSelect((event.currentTarget as HTMLSelectElement).value)}
-      >
-        <option value="all">{allLabel}</option>
-        {current.length > 1 && (
-          <option value="__multiple" disabled>
-            Multiple selected
-          </option>
-        )}
-        {options.map((option) => (
-          <option key={option} value={option}>
-            {format(option)}
-          </option>
-        ))}
-      </select>
-      {description && <p class="mt-1 hidden text-[11px] leading-snug text-[var(--bb-fg-muted)] sm:block">{description}</p>}
-    </label>
-  );
-}
-
-function SelectFilter({
-  label,
-  options,
-  current,
-  onSelect,
-  format,
-}: {
-  label: string;
-  options: string[];
-  current: string;
-  onSelect: (value: string) => void;
-  format: (value: string) => string;
-}) {
-  if (options.length === 0) return null;
-  return (
-    <label class="block">
-      <span class="mb-2 block text-xs font-semibold uppercase tracking-wide text-[var(--bb-fg-muted)]">{label}</span>
-      <select
-        aria-label={label}
-        class="h-9 w-full rounded-md border border-[var(--bb-border-default)] bg-[var(--bb-bg-primary)] px-3 text-sm font-medium text-[var(--bb-fg-primary)]"
-        value={current}
-        onChange={(event) => onSelect((event.currentTarget as HTMLSelectElement).value)}
-      >
-        {options.map((option) => (
-          <option key={option} value={option}>
-            {format(option)}
-          </option>
-        ))}
-      </select>
-    </label>
-  );
-}
-
-function CoverageSummary() {
-  let pickedCount = 0;
-  let compareHref: string | null = null;
-  let pickedIds: readonly string[] = [];
-  try {
-    const picking = usePickingState();
-    pickedCount = picking.pickedIds.length;
-    pickedIds = picking.pickedIds;
-    compareHref = picking.compareHref;
-  } catch {
-    // Unit tests may render Home without the provider; fall back to empty state.
-  }
-  const pickingLabel =
-    pickedCount === 0 ? "Compare →" : pickedCount === 1 ? "Compare 1 selected →" : `Compare ${pickedCount} selected →`;
-  // At single pick, compareHref is null (needs 2 for comparison), but builder still benefits from ?ids=<one> pinned entry.
-  if (!compareHref && pickedIds.length === 1) {
-    compareHref = `/results/compare?ids=${encodeURIComponent(pickedIds[0]!)}`;
-  }
-  return (
-    <div>
-      <div class="mb-2 text-xs font-semibold uppercase tracking-wide text-[var(--bb-fg-muted)]">Ranking scope</div>
-      <div class="flex flex-wrap gap-2">
-        <span class="rounded-full bg-[var(--bb-bg-elevated)] px-3 py-1.5 text-xs font-medium text-[var(--bb-fg-muted)]">
-          Ranked results only
-        </span>
-        {/* Compare handles an empty selection and sends readers to the shared
-            run finder, so this entry point needs no prefilled IDs. */}
-        <a
-          href={compareHref ?? "/results/compare/"}
-          data-testid="home-compare-entrypoint"
-          class="rounded-full border border-[var(--bb-border-default)] bg-[var(--bb-bg-primary)] px-3 py-1.5 text-xs font-medium text-[var(--bb-fg-primary)] no-underline shadow-sm hover:bg-[var(--bb-bg-elevated)]"
-        >
-          {pickingLabel}
-        </a>
-      </div>
+      <SkeletonBlock className="mt-8 h-64" />
     </div>
   );
 }
@@ -1340,48 +247,28 @@ function FlywheelStrip() {
   );
 }
 
-function SingleFilterGroup({
+type StatCardLabel = string | { singular: string; plural: string };
+
+function resolveStatCardLabel(value: number | string, label: StatCardLabel): string {
+  if (typeof label === "string") return label;
+  return typeof value === "number" && value === 1 ? label.singular : label.plural;
+}
+
+function StatCard({
+  value,
   label,
-  options,
-  current,
-  onSelect,
-  format,
-  optionTitle,
-  description,
+  detail,
 }: {
-  label: string;
-  options: string[];
-  current: string;
-  onSelect: (value: string) => void;
-  format: (value: string) => string;
-  optionTitle?: (value: string) => string;
-  description?: string;
+  value: number | string;
+  label: StatCardLabel;
+  detail?: string;
 }) {
-  if (options.length === 0) return null;
+  const resolvedLabel = resolveStatCardLabel(value, label);
   return (
-    <div>
-      <div class="mb-2 text-xs font-semibold uppercase tracking-wide text-[var(--bb-fg-muted)]">{label}</div>
-      <div class="flex flex-wrap gap-2">
-        {options.map((value) => {
-          const active = current === value;
-          return (
-            <button
-              key={value}
-              class={`rounded-full px-3 py-1.5 text-xs font-medium transition-colors ${
-                active
-                  ? "bg-[var(--bb-tone-info-bg)] text-[var(--bb-tone-info-fg)]"
-                  : "bg-[var(--bb-bg-elevated)] text-[var(--bb-fg-muted)] hover:bg-[var(--bb-border-default)]"
-              }`}
-              onClick={() => onSelect(value)}
-              aria-pressed={active}
-              title={optionTitle?.(value)}
-            >
-              {format(value)}
-            </button>
-          );
-        })}
-      </div>
-      {description && <p class="mt-1 text-[11px] leading-snug text-[var(--bb-fg-muted)]">{description}</p>}
+    <div class="card">
+      <div class="text-3xl font-bold text-[var(--bb-accent-hover)]">{value}</div>
+      <div class="mt-1 text-sm font-medium text-[var(--bb-data-fg-muted)]">{resolvedLabel}</div>
+      {detail && <div class="mt-2 text-xs text-[var(--bb-data-fg-subtle)]">{detail}</div>}
     </div>
   );
 }
