+++ results-explorer/src/pages/__tests__/fixtures/corpus.ts
@@ -1,0 +1,279 @@
+/**
+ * Shared corpus fixtures for the explorer's two index pages.
+ *
+ * The overview and the ranking table read the same three queries, so their
+ * suites read the same rows: a divergence between them would be a property of
+ * the fixtures rather than of the pages.
+ */
+
+export const TIMING_ELIGIBLE = {
+  has_display_timing: true,
+  valid_query_count: 2,
+  missing_query_count: 0,
+  zero_timing_count: 0,
+  display_exclusion_reason: null,
+  comparison_exclusion_reason: null,
+  ranking_exclusion_reason: null,
+};
+
+/**
+ * ResultRow fixtures - shape mirrors the explicit `bench.results` projection
+ * used by `listResults()`.
+ */
+export const RESULT_ROWS = [
+  {
+    result_id: "r1",
+    benchmark: "clickbench",
+    scale_factor: 0.1,
+    platform: "DuckDB",
+    platform_id: "duckdb",
+    driver_version: null,
+    run_date: "2026-04-17T12:00:00Z",
+    power_score: null,
+    total_duration_s: 12,
+    geomean_ms: 10,
+    display_geomean_ms: 10,
+    query_count: 2,
+    ...TIMING_ELIGIBLE,
+    trust_label: "maintainer-run",
+    visibility: "public-curated",
+    platform_version: null,
+    execution_mode: "sql",
+    tuning_mode: "tuned",
+    tuning_hash: null,
+    test_type: "power",
+    validation_status: "exact",
+    cost_usd: 1.1,
+    normalized_cost_usd: 1.1,
+    cost_status: "normalized",
+    cost_scope: "compute_only",
+    cost_model_version: "2026.05.0",
+    deployment_class: "cloud",
+    cloud_provider: "aws",
+    cloud_region: "us-east-1",
+    instance_or_warehouse: "MEDIUM",
+    warehouse_size: "MEDIUM",
+    storage_format: "parquet",
+    compliance_class: null,
+    is_ranking_eligible: true,
+    has_plans: false,
+    plans_published: false,
+    has_tuning: true,
+    bundle_download_url: "",
+  },
+  {
+    result_id: "r2",
+    benchmark: "clickbench",
+    scale_factor: 0.1,
+    platform: "SQLite",
+    platform_id: "sqlite",
+    driver_version: null,
+    run_date: "2026-04-17T12:00:00Z",
+    power_score: null,
+    total_duration_s: 24,
+    geomean_ms: 20,
+    display_geomean_ms: 20,
+    query_count: 2,
+    ...TIMING_ELIGIBLE,
+    trust_label: "community-submission",
+    visibility: "public-curated",
+    platform_version: null,
+    execution_mode: "sql",
+    tuning_mode: "auto",
+    tuning_hash: null,
+    test_type: "power",
+    validation_status: "exact",
+    cost_usd: 2.3,
+    normalized_cost_usd: null,
+    cost_status: "not_applicable_local",
+    cost_scope: null,
+    cost_model_version: null,
+    deployment_class: "local",
+    cloud_provider: null,
+    cloud_region: null,
+    instance_or_warehouse: null,
+    warehouse_size: null,
+    storage_format: null,
+    compliance_class: null,
+    is_ranking_eligible: true,
+    has_plans: false,
+    plans_published: false,
+    has_tuning: false,
+    bundle_download_url: "",
+  },
+  {
+    result_id: "r3",
+    benchmark: "tpch",
+    scale_factor: 1,
+    platform: "DuckDB",
+    platform_id: "duckdb",
+    driver_version: null,
+    run_date: "2026-04-16T12:00:00Z",
+    power_score: 3000,
+    total_duration_s: 60,
+    geomean_ms: 30,
+    display_geomean_ms: 30,
+    query_count: 22,
+    has_display_timing: true,
+    valid_query_count: 22,
+    missing_query_count: 0,
+    zero_timing_count: 0,
+    display_exclusion_reason: null,
+    comparison_exclusion_reason: null,
+    ranking_exclusion_reason: null,
+    trust_label: "maintainer-run",
+    visibility: "public-curated",
+    platform_version: null,
+    execution_mode: "sql",
+    tuning_mode: "tuned",
+    tuning_hash: null,
+    test_type: "power",
+    validation_status: "exact",
+    cost_usd: 5.5,
+    normalized_cost_usd: 5.5,
+    cost_status: "normalized",
+    cost_scope: "compute_only",
+    cost_model_version: "2026.05.0",
+    deployment_class: "cloud",
+    cloud_provider: "gcp",
+    cloud_region: "us-central1",
+    instance_or_warehouse: "LARGE",
+    warehouse_size: "LARGE",
+    storage_format: "parquet",
+    compliance_class: null,
+    is_ranking_eligible: true,
+    has_plans: false,
+    plans_published: false,
+    has_tuning: true,
+    bundle_download_url: "",
+  },
+  {
+    result_id: "r4",
+    benchmark: "star_schema",
+    scale_factor: 1,
+    platform: "Postgres",
+    platform_id: "postgres",
+    driver_version: null,
+    run_date: "2026-04-15T12:00:00Z",
+    power_score: null,
+    total_duration_s: 42,
+    geomean_ms: 21,
+    display_geomean_ms: 21,
+    query_count: 13,
+    ...TIMING_ELIGIBLE,
+    trust_label: "maintainer-run",
+    visibility: "public-curated",
+    platform_version: null,
+    execution_mode: "sql",
+    tuning_mode: null,
+    tuning_hash: null,
+    test_type: "power",
+    validation_status: "exact",
+    cost_usd: null,
+    normalized_cost_usd: null,
+    cost_status: "not_applicable_local",
+    cost_scope: null,
+    cost_model_version: null,
+    deployment_class: "local",
+    cloud_provider: null,
+    cloud_region: null,
+    instance_or_warehouse: null,
+    warehouse_size: null,
+    storage_format: null,
+    compliance_class: null,
+    is_ranking_eligible: true,
+    has_plans: false,
+    plans_published: false,
+    has_tuning: false,
+    bundle_download_url: "",
+  },
+];
+
+/** Per-platform summary rows - shape mirrors `bench.meta_leaderboard`. */
+export const META_LEADERBOARD_ROWS = [
+  { platform_id: "duckdb", platform: "DuckDB", avg_rank: 1, n_cohorts: 2 },
+  { platform_id: "sqlite", platform: "SQLite", avg_rank: 2, n_cohorts: 1 },
+];
+
+/**
+ * Per-variant cohort rows - shape mirrors `bench.cohort_metadata`.
+ * The pivot in `getMetaLeaderboardData` reconstructs the nested MetaLeaderboard
+ * from these rows.
+ */
+export const COHORT_ROWS = [
+  {
+    cohort_key: "clickbench-sf0.1-power",
+    benchmark: "clickbench",
+    scale_factor: 0.1,
+    phase: "power",
+    cohort_label: "ClickBench SF0.1",
+    cohort_href: "/results/clickbench/",
+    platform_count: 2,
+    cohort_ranked_count: 2,
+    cohort_ranking_exclusion_reason: null,
+    primary_metric: "display_geomean_ms",
+    primary_order: "asc",
+    platform_id: "duckdb",
+    platform: "DuckDB",
+    result_id: "r1",
+    short_id: "",
+    tuning_mode: "tuned",
+    trust_label: "maintainer-run",
+    ...TIMING_ELIGIBLE,
+    rank: 1,
+    metric_value: 10,
+    speedup_vs_best: 1,
+  },
+  {
+    cohort_key: "clickbench-sf0.1-power",
+    benchmark: "clickbench",
+    scale_factor: 0.1,
+    phase: "power",
+    cohort_label: "ClickBench SF0.1",
+    cohort_href: "/results/clickbench/",
+    platform_count: 2,
+    cohort_ranked_count: 2,
+    cohort_ranking_exclusion_reason: null,
+    primary_metric: "display_geomean_ms",
+    primary_order: "asc",
+    platform_id: "sqlite",
+    platform: "SQLite",
+    result_id: "r2",
+    short_id: "",
+    tuning_mode: "auto",
+    trust_label: "community-submission",
+    ...TIMING_ELIGIBLE,
+    rank: 2,
+    metric_value: 20,
+    speedup_vs_best: 0.5,
+  },
+  {
+    cohort_key: "tpch-sf1-power",
+    benchmark: "tpch",
+    scale_factor: 1,
+    phase: "power",
+    cohort_label: "TPC-H SF1",
+    cohort_href: "/results/tpch/",
+    platform_count: 1,
+    cohort_ranked_count: 1,
+    cohort_ranking_exclusion_reason: null,
+    primary_metric: "power_score",
+    primary_order: "desc",
+    platform_id: "duckdb",
+    platform: "DuckDB",
+    result_id: "r3",
+    short_id: "",
+    tuning_mode: "tuned",
+    trust_label: "maintainer-run",
+    has_display_timing: true,
+    valid_query_count: 22,
+    missing_query_count: 0,
+    zero_timing_count: 0,
+    display_exclusion_reason: null,
+    comparison_exclusion_reason: null,
+    ranking_exclusion_reason: null,
+    rank: 1,
+    metric_value: 3000,
+    speedup_vs_best: 1,
+  },
+];
